Step3 - 회차별 저장 및 확인
